PIC24/XC16 compiler does not define uintptr_t. Stop using uintptr_t, use int32_t instead.

+++ src/class.c
@@ -855,7 +855,7 @@ static void c_object_to_s(struct VM *vm, mrbc_value v[], int argc)
 	mrbc_printf_str( &pf, symid_to_str(v->instance->cls->sym_id), ' ' );
 	break;
       case 'x':
-	mrbc_printf_int( &pf, (uintptr_t)v->instance, 16 );
+	mrbc_printf_int( &pf, (uint32_t)v->instance, 16 );
 	break;
       }
     }
@@ -973,7 +973,7 @@ static void c_proc_to_s(struct VM *vm, mrbc_value v[], int argc)
 
   mrbc_printf_init( &pf, buf, sizeof(buf), "<#Proc:%08x>" );
   while( mrbc_printf_main( &pf ) > 0 ) {
-    mrbc_printf_int( &pf, (uintptr_t)v->proc, 16 );
+    mrbc_printf_int( &pf, (uint32_t)v->proc, 16 );
   }
   mrbc_printf_end( &pf );
